+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/VideoRecordDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface VideoRecordDao {
 
-    List<VideoRecord> queryVideoRecords(Integer aid, Integer last_count, Integer start_ts, Integer end_ts, Boolean limit,
+    List<VideoRecord> queryVideoRecords(Long aid, Integer last_count, Integer start_ts, Integer end_ts, Boolean limit,
                                         Integer offset, Integer ps);
 
-    Integer queryVideoRecordsCount(Integer aid, Integer start_ts, Integer end_ts);
+    Integer queryVideoRecordsCount(Long aid, Integer start_ts, Integer end_ts);
 
 }
