+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/VideoStaffExService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface VideoStaffExService {
 
-    List<VideoStaffEx> queryVideoStaffsByAid(Integer aid);
+    List<VideoStaffEx> queryVideoStaffsByAid(Long aid);
 }
