+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/VideoStaffService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface VideoStaffService {
 
-    List<VideoStaff> queryVideoStaffsByAid(Integer aid);
+    List<VideoStaff> queryVideoStaffsByAid(Long aid);
 
 }
