+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/user/UserFavoriteVideoService.java
@@ -7,16 +7,16 @@ import java.util.List;
 
 public interface UserFavoriteVideoService {
 
-    TddCommonResponse postUserFavoriteVideo(Long userid, Integer aid);
+    TddCommonResponse postUserFavoriteVideo(Long userid, Long aid);
 
-    TddCommonResponse deleteUserFavoriteVideo(Long userid, Integer aid);
+    TddCommonResponse deleteUserFavoriteVideo(Long userid, Long aid);
 
-    UserFavoriteVideo queryUserFavoriteVideo(Long userid, Integer aid);
+    UserFavoriteVideo queryUserFavoriteVideo(Long userid, Long aid);
 
-    Integer queryUserFavoriteVideoCount(Integer aid);
+    Integer queryUserFavoriteVideoCount(Long aid);
 
-    List<UserFavoriteVideo> queryUserFavoriteVideoUsers(Integer aid, Integer start_ts, Integer end_ts,
+    List<UserFavoriteVideo> queryUserFavoriteVideoUsers(Long aid, Integer start_ts, Integer end_ts,
                                                         String order_by, Integer desc, Integer pn, Integer ps);
 
-    Integer queryUserFavoriteVideoUsersCount(Integer aid, Integer start_ts, Integer end_ts);
+    Integer queryUserFavoriteVideoUsersCount(Long aid, Integer start_ts, Integer end_ts);
 }
